Changes to simplify customer concept
One customer class handles both external customers and internal employee
customers.
